Dramatic changes, now uses pop-build
This commit moves salt-bin to use the pop-build system
so that we only need to maintain one build system. This
also changes the nature of the salt-bin repo to be just
a collection of configs for pop-build to build Salt
with specific deps.
In effect, we are making Salt distributions